PO Note work

+++ notesService/Models/Notes.cs
@@ -14,9 +14,9 @@ namespace notesService.Models
         public int ref_id { get; set; }
         public int note_type_id { get; set; }
         public string note { get; set; }
-        public Double created_by { get; set; }
+        public int created_by { get; set; }
         public int updated_by { get; set; }
         public Double created_at { get; set; }
-        public int updated_at { get; set; }
+        public Double updated_at { get; set; }
     }
 }
